% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassPredictor.R
\name{ClassPredictorK}
\alias{ClassPredictorK}
\title{Class predictor of a K-Means Clustering using K-Nearest Neighbors}
\usage{
ClassPredictorK(train, test, k = 2, train_target, test_target)
}
\arguments{
\item{train}{A training data matrix}

\item{test}{A test data matrix}

\item{k}{Number of clusters for K-Means to be used with a default of 2 clusters}

\item{train_target}{A vector which is the target column from the training dataset}

\item{test_target}{A vector which is required to build a confusion matrix and accuracies}
}
\value{
\code{ClassPredictorK} returns an object of the class \code{'ClassPredictorK'} which has a vector of the class predicted, Confusion Matrix, Accuracy, False Positives and False Negatives. To see available attributes to pull type \code{names(lk)} where \code{lk} is an object of the class \code{ClassPredictorK}
}
\description{
Predicts target value in test data for a given training and test data
}
\examples{
#This example takes in age, height and weight and, targets as genders which are train[,5] and test[,5]
#and gives various predictions
train = read.csv('train.csv')
test = read.csv('test.csv')
lk = ClassPredictorK(train[,1:3], test[,1:3], k = 2, train[,5], test[,5])
names(lk)
}
