% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassPredictor.R
\name{ClassPredictorM}
\alias{ClassPredictorM}
\title{Class predictor of a Guassian Mixture Model using K-Nearest Neighbors}
\usage{
ClassPredictorM(train, test, k = 7, train_target, test_target)
}
\arguments{
\item{train}{A training data matrix}

\item{test}{A test data matrix}

\item{k}{Number of clusters for Guassian Mixture Model to be used with a default of 7 clusters}

\item{train_target}{A vector which is the target column from the training dataset}

\item{test_target}{A vector which is required to build a confusion matrix and accuracies
#' @return \code{ClassPredictorM} returns an object of the class \code{'ClassPredictorM'} which has a vector of the class predicted, Confusion Matrix, Accuracy, False Positives and False Negatives. To see available attributes to pull type \code{names(lm)} where \code{lm} is an object of the class \code{ClassPredictorM}}
}
\description{
Predicts target value in test data for a given training and test data
}
\examples{
#This example takes in age, height and weight and, targets as genders which are train[,5] and test[,5]
#and gives various predictions
train = read.csv('train.csv')
test = read.csv('test.csv')
lm = ClassPredictorM(train[,1:3], test[,1:3], k = 7, train[,5], test[,5])
names(lm)
}
