% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparer.R
\name{comparer}
\alias{comparer}
\title{Comparer}
\usage{
comparer(lk, lm)
}
\arguments{
\item{lk}{An object of class \code{ClassPredictorK} which can be built using a function in this package}

\item{lm}{An object of class \code{ClassPredictorM} which can be built using a function in this package}
}
\value{
Prints which model is better for the specific dataset working
}
\description{
Compares two objects of classes \code{ClassPredictorK} and \code{ClassPredictorM}
}
\examples{
train = read.csv('train.csv')
test = read.csv('test.csv')
lk = ClassPredictorK(train[,1:3], test[1:3], 2, train[,5],test[,5])
lm = ClassPredictorM(train[,1:3], test[,1:3],7, train[,5], test[,5])
comparer(lk, lm)
}
