% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalK.R
\name{optimalK}
\alias{optimalK}
\title{Optimal Number of Clusters}
\usage{
optimalK(data = data, till = 15)
}
\arguments{
\item{data}{A dataset for clustering}

\item{till}{A numeric value which tests from \code{2} to \code{till}}
}
\value{
optimalK returns a numeric value
}
\description{
Find optimal number of clusters for a data matrix using \code{kmeans} clustering and average silhouette coefficient
}
\examples{
#The example below takes in age, height and weight columns from the train data
#then tries values till 15 and gives an output of 2
train = read.csv('train.csv')
optimalK(data = train[,1:3], till = 15) #returns 2
}
