% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalK.R
\name{optimalKplot}
\alias{optimalKplot}
\title{Plot Optimal Number of Clusters}
\usage{
optimalKplot(data = data, FUNcluster = kmeans, method = c("silhouette"))
}
\arguments{
\item{data}{A dataset for clustering}

\item{FUNcluster}{The clustering algorithm which you want to use. Thee default is \code{kmeans}}

\item{method}{The method to find optimal clusters which may be silhouette, elbow or gap-statistic}
}
\value{
optimalKplot returns a plot visualization
}
\description{
Find optimal number of clusters for a data matrix using \code{kmeans} clustering and average silhouette coefficient
}
\examples{
#The example below takes in the age, height and weight and gives a plot based on average
#silhouette, suggesting 2 clusters
train = read.csv('train.csv')
optimalKplot(data = train[,1:3], kmeans, 'silhouette')
}
