% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plotter}
\alias{plotter}
\title{3D plotter}
\usage{
plotter(data = data, fill)
}
\arguments{
\item{data}{A dataset with 3 columns to plot}

\item{fill}{A vector by which data is to be color coded}
}
\value{
plotter returns a plot of class \code{plotly}
}
\description{
Plots a 3D plot on a data matrix as a plotly visualization
}
\examples{
#The below example is plotting the age, height and weight color coded by gender
train = read.csv('train.csv')
plotter(train[,1:3], fill = train[,5])
}
