% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plotterK}
\alias{plotterK}
\title{K-Means 3D plotter}
\usage{
plotterK(data, k)
}
\arguments{
\item{data}{A dataset with 3 columns to plot}

\item{k}{The number of clusters}
}
\value{
plotterK returns a plot of class \code{plotly}
}
\description{
Plots a 3D plot on a data matrix using K-Means Clustering as a plotly visualization
}
\examples{
#The below example is plotting the age, height and weight color coded by cluster
train = read.csv('train.csv')
plotterK(train[,1:3],2)
}
